squelch compiler warning
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@5906 8aca7d54-2c30-db11-9de9-000461428c89

+++ script/api/get_next.cc
@@ -5,7 +5,7 @@ SQInteger generic_get_next(HSQUIRRELVM vm, uint32 count)
 	SQInteger index;
 	if (SQ_SUCCEEDED(sq_getinteger(vm, -1, &index))) {
 		// now increase index
-		if (0<=index  &&  index+1<count) {
+		if (0<=index  &&  (uint32)index+1<count) {
 			index ++;
 			sq_pushinteger(vm, index);
 		}
